Testing Pull requests into main.

+++ channel_templates/up.sh
@@ -8,34 +8,39 @@ until docker ps | grep -q bitcoind; do
     sleep 0.1;
 done;
 
-# set these funcs and vars here for testing
-lncli() {
-    "./../lightning-cli.sh" "$@"
-}
+RETAIN_CACHE=false
 
-bcli() {
-    "./../bitcoin-cli.sh" "$@"
-}
-
-export -f lncli
-export -f bcli
-
-
-# clear out the node addrs and pubkey cache
-rm -f ./node_addrs.txt
-rm -f ./node_pubkeys.txt
-
-# cache node pubkeys
-for ((NODE_ID=0; NODE_ID<CLN_COUNT;NODE_ID++)); do
-    pubkey=$(lncli --id=$NODE_ID getinfo | jq -r ".id")
-    echo "$pubkey" >> node_pubkeys.txt
+for i in "$@"; do
+    case $i in
+        --retain-cache=*)
+            RETAIN_CACHE="${i#*=}"
+            shift
+        ;;
+        *)
+        ;;
+    esac
 done
 
-# cache node addrs
-for ((NODE_ID=0; NODE_ID<CLN_COUNT;NODE_ID++)); do
-    addr=$(lncli --id=$NODE_ID newaddr | jq -r ".bech32")
-    echo "$addr" >> node_addrs.txt
-done
+# recache node addrs and pubkeys if not specified otherwise
+if [ "$RETAIN_CACHE" == false ]; then
+    echo "Caching node info..."
+
+    rm -f ./node_addrs.txt
+    rm -f ./node_pubkeys.txt
+
+    for ((NODE_ID=0; NODE_ID<CLN_COUNT;NODE_ID++)); do
+        pubkey=$(lncli --id=$NODE_ID getinfo | jq -r ".id")
+        echo "$pubkey" >> node_pubkeys.txt
+    done
+    echo "Node pubkeys cached"
+
+    for ((NODE_ID=0; NODE_ID<CLN_COUNT;NODE_ID++)); do
+        addr=$(lncli --id=$NODE_ID newaddr | jq -r ".bech32")
+        echo "$addr" >> node_addrs.txt
+    done
+        echo "Node addresses cached"
+
+fi
 
 ./bitcoind_load_onchain.sh
 
@@ -45,6 +50,9 @@ if [ "$BTC_CHAIN" = regtest ]; then
     ./cln_load_onchain.sh
 
     ./bootstrap_p2p.sh
+    TIME_PER_CLN_NODE=3
+    
+    sleep $((CLN_COUNT * TIME_PER_CLN_NODE))
 
     ./regtest_prism.sh
